reference/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.13.4
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.0
defusedxml==0.7.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.2.1
django-crontab==0.7.1
django-storages==1.14
django-filter==25.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-recaptcha==4.1.0
django-stubs-ext==5.2.0
django-taggit==6.1.0
django-tasks==0.7.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
django_components==0.139.1
djangorestframework==3.16.0
djc_core_html_parser==1.0.2
dotenv==0.9.9
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filelock==3.18.0
filetype==1.2.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-storage==3.1
googleapis-common-protos==1.70.0
gunicorn==23.0.0
h11==0.16.0
httplib2==0.22.0
icalendar==5.0.11
identify==2.6.10
idna==3.10
laces==0.1.2
nodeenv==1.9.1
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
pillow_heif==0.21.0
platformdirs==4.3.8
pre_commit==4.2.0
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pydantic==2.11.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uv==0.7.3
uvicorn==0.34.0
virtualenv==20.31.2
wagtail==7.0
whitenoise==6.9.0
Willow==1.10.0
